requirements.txt
# Deep Learning Framework
#torch>=2.0.0
#torchvision>=0.15.0

# Spiking Neural Networks
snntorch>=0.7.0

# CLIP for Stage 2
ftfy>=6.1.0
regex>=2023.0.0
git+https://github.com/openai/CLIP.git

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Data & Config
pyyaml>=6.0
h5py>=3.9.0
python-dotenv>=1.0.0

# Training & Logging
tensorboard>=2.13.0
tqdm>=4.65.0

# Evaluation Metrics
scikit-image>=0.21.0
lpips>=0.1.4

# Visualization
matplotlib>=3.7.0
